Move sanitize_dictionary to misc.test_utils
